refactor: move cn fn to utils dir
