+++ components.json
@@ -11,7 +11,6 @@
     "prefix": ""
   },
   "aliases": {
-    "components": "@/components",
-    "utils": "@/lib/utils"
+    "components": "@/components"
   }
 }
