+++ src/components/ui/command.tsx
@@ -6,7 +6,7 @@ import { Command as CommandPrimitive } from 'cmdk';
 import * as React from 'react';
 
 import { Dialog, DialogContent } from '@/components/ui/dialog';
-import { cn } from '@/lib/utils';
+import { cn } from '@/utils/cn';
 
 const Command = React.forwardRef<
   React.ElementRef<typeof CommandPrimitive>,
