+++ src/components/wallet/SendAddressForm.tsx
@@ -36,9 +36,9 @@ import {
   PayLiquidAddressMutationVariables,
 } from '@/graphql/mutations/__generated__/pay.generated';
 import { useWalletInfo } from '@/hooks/wallet';
-import { cn } from '@/lib/utils';
 import { useKeyStore } from '@/stores/keys';
 import { toWithError } from '@/utils/async';
+import { cn } from '@/utils/cn';
 import { handleApolloError } from '@/utils/error';
 import { cryptoToUsd } from '@/utils/fiat';
 import {
